+++ src/main/kotlin/com/capston/sumnote/note/service/NoteServiceImpl.kt
@@ -19,62 +19,42 @@ class NoteServiceImpl(
     private val notePageRepository: NotePageRepository
 ) : NoteService {
 
+    /**
+     * 노트 생성
+     */
     @Transactional
     override fun createNote(dto: CreateNoteDto, email: String): CustomApiResponse<*> {
-
-        // 이메일로 멤버 찾기
         val member = memberRepository.findByEmail(email).orElse(null)
-            ?: return CustomApiResponse.createFailWithoutData(404, "Member not found")
+            ?: return CustomApiResponse.createFailWithoutData(HttpStatus.NOT_FOUND.value(), "사용자를 찾을 수 없습니다.")
 
-        // 노트 생성과 저장
         val note = Note(title = dto.note.title, member = member)
         val savedNote = noteRepository.save(note)
-
-        // 노트 페이지 생성과 저장
         dto.notePages.forEach {
-            val notePage = NotePage(
-                title = it.title,
-                content = it.content,
-                note = savedNote
-            )
-            notePageRepository.save(notePage)
+            notePageRepository.save(NotePage(title = it.title, content = it.content, note = savedNote))
         }
-
-        // 응답
-        return CustomApiResponse.createSuccessWithoutData<Unit>(201, "노트가 정상적으로 생성되었습니다.")
+        return CustomApiResponse.createSuccessWithoutData<Unit>(HttpStatus.CREATED.value(), "노트가 정상적으로 생성되었습니다.")
     }
 
-    override fun findRecentNotesLimited(email: String): CustomApiResponse<List<GetNotesDto>> {
-        val notes = noteRepository.findTop5ByMemberEmailOrderByLastModifiedAtDesc(email)
-        return CustomApiResponse.createSuccess(HttpStatus.OK.value(), notes, "최근 생성된 노트 5개 조회에 성공하였습니다.")
+    /**
+    * 모든 노트 조회
+     */
+    override fun findNotesByType(email: String, type: String): CustomApiResponse<*> = when (type) {
+        "home" -> CustomApiResponse.createSuccess(HttpStatus.OK.value(), noteRepository.findTop5ByMemberEmailOrderByLastModifiedAtDesc(email), "최근 생성된 노트 5개 조회에 성공하였습니다.")
+        "all" -> CustomApiResponse.createSuccess(HttpStatus.OK.value(), noteRepository.findAllByMemberEmailOrderByLastModifiedAtDesc(email), "모든 노트 조회에 성공하였습니다.")
+        else -> CustomApiResponse.createFailWithoutData(HttpStatus.BAD_REQUEST.value(), "type은 home 또는 all 이어야 합니다.")
     }
 
-    override fun findAllNotesSorted(email: String): CustomApiResponse<List<GetNotesDto>> {
-        val notes = noteRepository.findAllByMemberEmailOrderByLastModifiedAtDesc(email)
-        return CustomApiResponse.createSuccess(HttpStatus.OK.value(), notes, "모든 노트 조회에 성공하였습니다.")
-    }
-
-    override fun getNote(noteId: Long): CustomApiResponse<*> {
-
-        // 노트 찾기 + 예외 처리
+    /**
+     * 노트 내용 조회
+     */
+    override fun getNote(email: String, noteId: Long): CustomApiResponse<*> {
         val note = noteRepository.findById(noteId).orElse(null)
             ?: return CustomApiResponse.createFailWithoutData(HttpStatus.NOT_FOUND.value(), "id가 " + noteId + "인 노트는 존재하지 않습니다.")
-
-        // 응답 데이터 만들기
         val noteDetail = NoteDetail(noteId = note.id!!, title = note.title!!)
         val notePagesDetails = note.notePages.map {
-            NotePageDetail(
-                notePageId = it.id!!,
-                title = it.title!!,
-                content = it.content!!,
-                isQuizExist = it.isQuizExists ?: false
-            )
+            NotePageDetail(notePageId = it.id!!, title = it.title!!, content = it.content!!, isQuizExist = it.isQuizExists ?: false)
         }
-
-        val getNoteDto = GetNoteDto(note = noteDetail, notePages = notePagesDetails)
-
-        // 응답
-        return CustomApiResponse.createSuccess(HttpStatus.OK.value(), getNoteDto, "노트 조회에 성공하였습니다.")
+        return CustomApiResponse.createSuccess(HttpStatus.OK.value(), GetNoteDto(note = noteDetail, notePages = notePagesDetails), "노트 조회에 성공하였습니다.")
     }
